Better type choice.
git-svn-id: svn://svn.code.sf.net/p/chicken/code/trunk/cotvnc@767 9ce1af06-8ae8-4ba5-be20-3b57e42882c5

+++ Source/SshTunnel.m
@@ -483,7 +483,7 @@ static BOOL portUsed[TUNNEL_PORT_END - TUNNEL_PORT_START];
         errno = 0;
     } else {
         const char  *cStr = [str UTF8String];
-        int         len = strlen(cStr);
+        size_t      len = strlen(cStr);
         char        *buf = malloc(len + 1);
 
         memcpy(buf, cStr, len);
